Introduce capability layer.
This change introduces a capability layer with negotiation protocol.

The current implementation supports encryption capability using ECDHE (Ephemeral X25519) with AEAD for authenticated encryption.

This makes supporting TLS something of the past, because it turned out to be a nightmare in terms of implementation and performance.

The current capability protocol has been integrated in "basic" form in the current client and server implementations.

The capability protocol is, more or less, a prototype
that avoids speculative over-design, and might undergo many changes.