Update names and documents.
This change updates the names and naming scheme and improves the encryption APIs to reduce system calls.

This and subsequent changes stop using Stream or I/O to refer to transport layer types, using simply "Transport" instead.

The subsequent changes will refine the capability layer and introduce an identity verification mechanism that is not based on x.509 mechanisms and CA. It will be based on a protocol **similar** to OPAQUE protocol, where a client-provided token is used as an identification token and servers only store PRF-based verifiers without storing secrets or tokens themselves.

Unlike TLS, the capability layer will not evolve to manage its own messages and state and will remain a simple negotiation mechanism with core primitives. The RPC frame and application message, however, will be updated to support system messages. These allow session/connection handlers to react to state changes and update active sessions on the fly without having to coordinate with other layers (like the record layer in TLS, is_handshaking?... blah blah blah, with countless checks and stored flags just to determine the state of the record layer!).

Thus, the nonsense and massive inefficiency of TLS are completely avoided, and certificates and CAs will not be considered for identity verification.