Introduce new synchronization primitive.
This change introduces a new synchronization primitive `ReleaseBarrier`.

The main motivation for this is the need for lightweight and safe cancellation mechanism, where cancellation is performed gracefully based on **passive consensus**.

The design and usage of this barrier as cancellation mechanism is based on the following requirements:

- Using existing mechanism: tokio's abort handle of the spawned task we have already paid the cost for.

- No complex abort chain with multiple scheduled futures.

- Lightweight with minimal overhead: No complex chain of repeated checks and linked lists.

- Targeted with region-based control: The region has the power to prevent cancellation, with ability to know if it has stuck, so it can rollback changes.

So this barrier offers a simple solution: The system tries to execute cancellation logic, if, and only if no objection is visible, otherwise the cancellation call will be awaited.

The objection to cancellation is expressed by acquiring ref-counted locks in critical sections/regions, these locks prevent the barrier from releasing.

The objection is resolved by dropping the lock,  so unless all locks are released, the barrier can't resolve and must wait.

With this approach the overhead is very small, tied to specific regions, and the only future scheduled and polled is the cancellation future.

You could argue that this design requires waiting for indefinite amount of time, but I can argue that while this is true, it can be avoided  where needed, after all it is called "graceful" not "forceful" cancellation, if not awaited structurally, it must be awaited detached, making sure the process doesn't exit before it is done.

Side note: If releasing is taking too long to complete, it is a good sign of performance issues in those regions anyway.

Things may evolve differently, with changes and refinements, but I think the approach in general is very simple and clean for the problem of safe cancellation of detached tasks, it keeps the code lean and clean, especially the loops.