Rename ReleaseBarrier.
This changes renames the barrier-thing to something close to its original name `Latch`.

This was the name at the time of birth, but since std::latch in C++ doesn't support incrementing with fixed max count, I decided to change the name to avoid confusion, but the name I came up with was not less confusing than the original name, so it is better to settle this with some Apple-style naming scheme (Dynamic Island-something) to make the swift team happy, because I feel they aren't these days.