Add connection management subsystem.
This change introduces a connection management subsystem to aid connections' tracking and support planned shutdown of the server.

The added machinery adds overhead to creating new connections, but this overhead could be the lowest under the sun, if we consider the mission to accomplish.

The high efficiency of the design compared to what it does comes from eliminating heap allocations to store connection state, and avoiding off-the-shelf abstractions usually used in such contexts.

The current design relies heavily on unsafe code to cut corners which introduces complexity and concurrency bugs if not refactored carefully.

It is worth noting that this is the first iteration that will further refined and optimized. It could also contain bugs, because concurrency makes unsafe code several orders of magnitudes harder.

It is also worth noting in that regard that everything I did is based on the available documentations which are very fragmented and opinionated.

As final thought, most people used to use "managed" servers, where the management overhead is just accepted, but I will extend the current implementation later with type state with different wiring, that starts unmanaged connections for use cases where managing connections and planned shutdown are not useful, so the performance penalty of tracking could be avoided.