Refactor connection functions.
This change "unbox" keys again, and does some code cleanups.

The concerns about the size of keys is valid, but the I think boxing keys was a quick decision based on some encountered stack overflow issues.

Anyway I will not be quick to judge things, but this setup seems much faster and lighter on average, and I will keep it this way until I know better.