Remove bidirectional calls support in client impl.
This is rather tactical than strategic decision, I think bidirectional calls are rather useful in certain cases, but the current implementation can't be fully functional without server support, and the current server doesn't support that yet.

The reason for delaying support for bidirectional calls in server is that it needs a matching machinery per session to route replies, because the channel/stream is designed for long running interactive sessions, where messages are received in an arbitrary order. But bidirectional calls could be the rare case, so we need a better design, where bidirectional calls are "blocking", this is similar to the stream/group design I am working on, where no in-between unrelated messages are expected/allowed.

Anyway, forcing each server session to have "matcher" "just-in-case" does't align with the elegance and efficiency goals of this project.