Introduce calling context.
This change adds calling context to service calls and brings back some sort of "bidirectional" calls, it also deprecates the notification message.

These changes need to be addressed one by one:

- Context: a low-overhead struct that has two pointers/references to the message id and the sender that exposes methods for sending and calling within the service call method. This allows the service to optionally send replies and enables better scalability of the API in general.

- Notification removal: notification was simply implemented as  a normal call that doesn't requires reply, and you know upfront that the target method doesn't return. For the sake of simplicity, it is up to the service to define which method returns and which one doesn't, especially that the service trait doesn't require returning reply anymore.

- The so called "bidirectional" calls: Bidirectional calls within the original design is that the server can call a method on the connected peer besides the usual call-wait/await-reply cycles. The problem is that this useful feature can introduce overhead on the server side for practically "rare" use cases, a thing I am unwilling to accept, at the same time wanting to have such feature as part of the core design, not as patchwork later.

So in order to support this as cheaply as like nearly nothing, the call-wait/await-reply cycle is replaced with call-callback without waiting/awaiting. So the reply, if necessary, can be just callback to specialized method with or without data.

Both client and server can opt-out of supporting bidirectional calls without paying storage and processing cost for it.

That leaves us with last message type to be added, "Sequence", AKA "Stream", but that for another day.