Introduce IORing.
This change introduces IORing.

IORing is a concurrent ring buffer that acts as a highly optimized  bounded communication channel.

The current implementation is considered unstable, but it will get close eyes in the next iterations and it will be further refined and improved.

IORing is going the be the "backbone" of many synchronizations and in-process communication pathways.