Introduce the Report API.
This change introduce the Report interface, a trait that act as universal API for reporting agents.

It introduce also a complete implementation of the API that uses stdio of the process as its output.

Please take note that both the trait and the implementation are considered unstable, and will be refined further with more usage.

The async version of the report API has been implemented many times and removed and finally I have abandoned the idea of having one.