Refactor reporter.
I think I overestimated the time needed to get the IORing done completely right, because currently it has many flaws, and relies on tricks like free capacity to remain functional.

So for the sake of keeping things functional without flaws, I will remove it for now, and use the lock if stdio directly like it was planned, before merging the ring, in the hope that I can get it done right without flaws within bounded and small time frame.