Refactor IORing.
The previous tweak was incorrect. Sorry for the oversight.

Anyway a better "garbage collection" strategy is needed.
 Meanwhile, only consumer can reclaim segments.

Keep in mind that this thing is still experimental, and was never planned until recently, I never intended to go to the business of implementing ring buffers, but the cancellation safety and other optimizations were a sort of a "push", but my head is not in the mood of doing complex concurrent data structures, it is just I am forced to.

In case you are wondering, I know that both std and tokio have channels, but they operate on structure and owned containers, not on bytes, anyway, I can't outsource this.