Add receive method.
This change adds a new method "receive" that returns a segment with published data, and removes the "receive_into" method.

This enables consumer to use the same buffer as "work bench" for further processing and routing to/offloading into another stream, regardless of whether the stream is sync or async.