Update names.
This change renames some variables and updates the coding style.

I must admit that this project could be very annoying to "String" heads and some people of Arc<Mutex<T>>, Box Box Box, dyn X and hidden control flow behind "cryptic" macros at every corner, with errors that try to tell stories instead of simply reporting a failure-point with minimal overhead.

Nevertheless, my goal is not to annoy anybody, it is just that I am trying to do the opposite of all of this, because I am just fed up with it all.