Refactor transport traits.
This change introduces custom IO traits, that form the basis of building transport layers.

This change is part of the policy, where the exposed/required interfaces/traits must be native to this project or part of standard library.

Since the standard library lacks the async version of the I/O traits they must be provided by this project.

The point of these traits at the moment is not to provide "special" functionality, but to use native traits as abstraction over a core component, although special things might be added later.