Refactor data-flow pathways.
This change is part of series of design-changes that aim at refining the architecture to be more of "data-flow" oriented, in sense that entire system is seen as a set of data lines, pathways directed and redirected toward/by entities that may apply transformation to the stream/data-flow, or try to construct components from its content.

This extends the design to be much more flexible and eventually scalable, and improves reasoning about the system in general, it also sets the door wide open for wide variety of optimizations that can radically improve performance.

These changes are still not fully refined, and carry some old artifacts and some duplicated fragments but eventually things will get better shape.