Centralize encoding and decoding behavior.
This change removes encoding and decoding methods from header and error in order to make "Message" the encoding and decoding "engine" of "everything", with optimized methods for many cases. These special cases are still ignored with a lot of space for optimizations in terms of allocations and runtime checking in many areas.