Refactor directive and update docs.
This change renames reply to return, and update directive's docs and some other docs.

This is again part of the paradigm shift of the project, and probably the last move as a departure from the Http-like terminology.

This project aims beyond the era of "bad jokes" like Http and the web, where a serious "systems" connect the world instead of these immature cheap hacks, that for a strange reason took over remote APIs space.

Meanwhile I wish WASM all the best, as it is the only thing that could make this pile of garbage "the web" a sort of "system" little bit  and help humanity leave the dark ages of HTML/JS/TS...behind.

The sun is rising...