Refactor MessageID.
This change refactors MessageID to be simple integer instead of UUID and refactors matching and publishing pathways in client.

This change introduces also "MatchingUnit" which does direct matching by using direct indexing.

These changes are motivated by streamlining and improving the matching logic by means of:

- Faster generation of IDs.

- Removing associative search (hashing).

- Transforming macro-synchronization into micro-synchronization (per slot instead of per collection).

There are several challenges to this design like correct indexing and ownership and limited capacity, but this is the right direction overall.