Refactor frame data.
This change removes unesssary cells.

This design will be revised to be more robust in terms of memory safety, because Rust's definition of memory-safety does not consider forgetting a type (not running its destructor) as memory-safety hazard, and since "futures" are "leaky abstractions", there is no point of documenting something like "the returned future must be dropped after being polled", so either the entire call should marked as "unsafe", or things should take different shape to satisfy this formal definition of "memory-safety".

But I am already planning major refactoring based on "pooling" to satisfy one of my major design goals which is to utilize static reusable memory allocations as much as possible.