Introduce IOPool.
This change introduces IOPool, a pool of pre-allocated memory regions for I/O operations.

This change and subsequent changes are part of series of changes that aim at improving the design to be centered around memory, in sense that memory is treated as fallible "I/O resource", that requires "strict" management.

Another motivation of these changes is that nothing in this project is designed to handle panic. Panic is not part of the control flow, and it is strictly a "fatal event", AKA terminating-error, nothing should survive panic, and trying to allocate dynamically should be a fallible case, where proper "safe-landing" can take place instead of just "panic".

Anyway, the pool's 'implementation is sort of "enabler" for further refactoring, but may be changed in the future.