Introduce IOSegment trait.
This change introduce  IOSegment trait, a trait that generalizes over allocated memory regions with common methods used to write and read untyped data.

This change is part of series of changes aiming at:

1 - Taming and controlling memory usage in a way helps constructing clear picture about the data lifecycle and memory usage in general.

2 - Constraining dynamic allocation for better performance, especially for applications where sizes are strictly defined as part of the application API definition and documentation and the "open bound" is not required.

3 - Enabling zero-copy encoding and decoding in safe and "sane" manner by more specialized types with on-demand explicit clone behavior.

4 - Keeping the door open for unknown sizes, or as "lazy" option for folks who don't want to "count their bytes".

All of the above is not an "easy ride" to implement, especially is "sane" way, where complexity is well managed and exposed via elegant and efficient APIs, but it just seems the right direction.