Refactor MessageStore.
This change refactors MessageStore to implement IOSegment trait and updates the adapter logic.

This is a step for a transition to "swappable" segments in order to "lease" messages with borrowed decoders instead of copying, because copying from and to is more of a concern than dynamic allocations, this is because general purpose allocators while not especially optimized pieces of software, but allocating of alike-sizes is likely to return a retained pre-allocated chunk in many cases.

Anyway, there is a way ahead to craft a highly optimized "circuits", but I can easily imagine a future where designing a software is centered around designing its memory "subsystem" instead of churning general purpose allocators with those "malloc" in random pieces at random places, in the same way I can imagine a future where everything "IPC" can be done via "protocoled" shared state in user-mode, instead of "poking" the kernel for every damn invocation that operates on that state with a ton of copying data between these scattered "buffers".