Update project's structure.
This change updates the projects structure and makes the current implementation in rust independent crate.

This current implementation will not retain the current encoding/decoding logic, which will be replaced by the new implementation of the "ABI" and the runtime-support types that will be used by the code-generator. Other things will remain, and more will be added, especially other IPC-mechanisms with different transport models.

The root crate will host the reference implementation of the compiler and its code-generators backends.