Refactor project structure.
This change revert making separate crate for the runtime-support.

The current goal is to define the data model and to describe services in terms of types defined in the data model and interfaces instead of communicating memory-layouts.

The compiler and code-gen approach just serves "automation", but all types and components can be composed manually also.

It is too early to think in terms of "other-languages", especially that..I am not that interested.