Add specification document.
This change introduces the specification draft, which describes the "canonical" data model and how it is going to be expressed in the definition language.

This specification is still a draft, and its implementation is not yet reflected in current implementation.

The new conforming implementation will be rolled out little by little to replace the bits and pieces of the current implementation without mess.

The compiler and the code generator will be introduced as last step, after making sure that the composition model is well implemented and working without issues.