Introduce the memory module.
This change introduces a core module that defines the primitive types and the core traits for reading and writing to memory in the "new implementation".

The current plan is to introduce stable stuff as "parallel" implementation, and doing staged transition using these types in "manual fashion". This will give a better picture for aiding compiler development.

More stuff will show up when I feel they are unlikely to undergo major changes.

It is worth noting that correct implementation of the definition language and its compiler may take a while, especially that the generated code will not depend on the layout and access rules of the target language, so the correct storage layout and correct composition of things depend on the compiler doing correct things when translating abstract definitions to memory layouts accessed and written by relative offsets.

Translating parsed definitions to correct storage layouts is what I am currently working on.